+++ age/src/protocol.rs
@@ -52,14 +52,6 @@ pub trait Callbacks {
     fn request_passphrase(&self, description: &str) -> Option<SecretString>;
 }
 
-struct NoCallbacks;
-
-impl Callbacks for NoCallbacks {
-    fn request_passphrase(&self, _description: &str) -> Option<SecretString> {
-        None
-    }
-}
-
 /// Handles the various types of age encryption.
 enum EncryptorType {
     /// Encryption to a list of recipients identified by keys.
@@ -227,9 +219,10 @@ impl<R: AsyncRead + Unpin> Decryptor<R> {
 mod tests {
     use secrecy::SecretString;
     use std::io::{BufReader, Read, Write};
+    use std::iter;
 
     use super::{Decryptor, Encryptor};
-    use crate::keys::{Identity, RecipientKey};
+    use crate::{identity::IdentityFile, keys::RecipientKey, Identity};
 
     #[cfg(feature = "async")]
     use futures::{
@@ -241,7 +234,10 @@ mod tests {
     #[cfg(feature = "async")]
     use futures_test::task::noop_context;
 
-    fn recipient_round_trip(recipients: Vec<RecipientKey>, identities: &[Identity]) {
+    fn recipient_round_trip(
+        recipients: Vec<RecipientKey>,
+        identities: impl Iterator<Item = Box<dyn Identity>>,
+    ) {
         let test_msg = b"This is a test message. For testing.";
 
         let mut encrypted = vec![];
@@ -264,7 +260,10 @@ mod tests {
     }
 
     #[cfg(feature = "async")]
-    fn recipient_async_round_trip(recipients: Vec<RecipientKey>, identities: &[Identity]) {
+    fn recipient_async_round_trip(
+        recipients: Vec<RecipientKey>,
+        identities: impl Iterator<Item = Box<dyn Identity>>,
+    ) {
         let test_msg = b"This is a test message. For testing.";
         let mut cx = noop_context();
 
@@ -341,18 +340,18 @@ mod tests {
     #[test]
     fn x25519_round_trip() {
         let buf = BufReader::new(crate::keys::tests::TEST_SK.as_bytes());
-        let sk = Identity::from_buffer(buf).unwrap();
+        let sk = IdentityFile::from_buffer(buf).unwrap();
         let pk: RecipientKey = crate::keys::tests::TEST_PK.parse().unwrap();
-        recipient_round_trip(vec![pk], &sk);
+        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 
     #[cfg(feature = "async")]
     #[test]
     fn x25519_async_round_trip() {
         let buf = BufReader::new(crate::keys::tests::TEST_SK.as_bytes());
-        let sk = Identity::from_buffer(buf).unwrap();
+        let sk = IdentityFile::from_buffer(buf).unwrap();
         let pk: RecipientKey = crate::keys::tests::TEST_PK.parse().unwrap();
-        recipient_async_round_trip(vec![pk], &sk);
+        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 
     #[test]
@@ -387,7 +386,7 @@ mod tests {
         let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
             .parse()
             .unwrap();
-        recipient_round_trip(vec![pk], &[sk.into()]);
+        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 
     #[cfg(feature = "async")]
@@ -398,7 +397,7 @@ mod tests {
         let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_RSA_PK
             .parse()
             .unwrap();
-        recipient_async_round_trip(vec![pk], &[sk.into()]);
+        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 
     #[test]
@@ -408,7 +407,7 @@ mod tests {
         let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
             .parse()
             .unwrap();
-        recipient_round_trip(vec![pk], &[sk.into()]);
+        recipient_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 
     #[cfg(feature = "async")]
@@ -419,6 +418,6 @@ mod tests {
         let pk: RecipientKey = crate::ssh::recipient::tests::TEST_SSH_ED25519_PK
             .parse()
             .unwrap();
-        recipient_async_round_trip(vec![pk], &[sk.into()]);
+        recipient_async_round_trip(vec![pk], iter::once(Box::new(sk) as Box<dyn Identity>));
     }
 }